[Decomp][Creature] Cohesion split: extract spell cooldowns to CreatureSpellCooldown.cpp
