+++ src/game/Object/Creature.cpp
@@ -2894,95 +2894,11 @@ Unit* Creature::SelectAttackingTarget(AttackingTarget target, uint32 position, S
     return NULL;
 }
 
-/**
- * @brief Stores an absolute cooldown end time for a creature spell.
- *
- * @param spell_id The spell identifier.
- * @param end_time The cooldown end time.
- */
-void Creature::_AddCreatureSpellCooldown(uint32 spell_id, time_t end_time)
-{
-    m_CreatureSpellCooldowns[spell_id] = end_time;
-}
 
-/**
- * @brief Stores the application time for a spell category cooldown.
- *
- * @param category The spell category.
- * @param apply_time The time when the category cooldown started.
- */
-void Creature::_AddCreatureCategoryCooldown(uint32 category, time_t apply_time)
-{
-    m_CreatureCategoryCooldowns[category] = apply_time;
-}
 
-/**
- * @brief Adds cooldown tracking for a creature spell and its category.
- *
- * @param spellid The spell identifier.
- */
-void Creature::AddCreatureSpellCooldown(uint32 spellid)
-{
-    SpellEntry const* spellInfo = sSpellStore.LookupEntry(spellid);
-    if (!spellInfo)
-    {
-        return;
-    }
 
-    uint32 cooldown = GetSpellRecoveryTime(spellInfo);
-    if (cooldown)
-    {
-        _AddCreatureSpellCooldown(spellid, time(NULL) + cooldown / IN_MILLISECONDS);
-    }
 
-    if (spellInfo->Category)
-    {
-        _AddCreatureCategoryCooldown(spellInfo->Category, time(NULL));
-    }
-}
 
-/**
- * @brief Checks whether a spell category cooldown is still active.
- *
- * @param spell_id The spell identifier.
- * @return true if the category cooldown is active; otherwise, false.
- */
-bool Creature::HasCategoryCooldown(uint32 spell_id) const
-{
-    SpellEntry const* spellInfo = sSpellStore.LookupEntry(spell_id);
-    if (!spellInfo)
-    {
-        return false;
-    }
-
-    CreatureSpellCooldowns::const_iterator itr = m_CreatureCategoryCooldowns.find(spellInfo->Category);
-    return (itr != m_CreatureCategoryCooldowns.end() && time_t(itr->second + (spellInfo->CategoryRecoveryTime / IN_MILLISECONDS)) > time(NULL));
-}
-
-/**
- * @brief Gets the remaining cooldown delay for a creature spell.
- *
- * @param spellId The spell identifier.
- * @return Remaining cooldown in seconds.
- */
-uint32 Creature::GetCreatureSpellCooldownDelay(uint32 spellId) const
-{
-    CreatureSpellCooldowns::const_iterator itr = m_CreatureSpellCooldowns.find(spellId);
-    time_t t = time(NULL);
-    return uint32(itr != m_CreatureSpellCooldowns.end() && itr->second > t ? itr->second - t : 0);
-}
-
-/**
- * @brief Checks whether a spell or its category is currently on cooldown.
- *
- * @param spell_id The spell identifier.
- * @return true if a cooldown is active; otherwise, false.
- */
-bool Creature::HasSpellCooldown(uint32 spell_id) const
-{
-    CreatureSpellCooldowns::const_iterator itr = m_CreatureSpellCooldowns.find(spell_id);
-    return (itr != m_CreatureSpellCooldowns.end() && itr->second > time(NULL)) || HasCategoryCooldown(spell_id);
-}
 
 /**
  * @brief Checks whether the creature is currently evading back home.
